feat: Generate customer survey data and save as CSV file
This commit adds code to generate customer survey data and saves it as a CSV file in the 'market analysis' directory. The generated data includes information about the customer's favorite drink, favorite food, favorite dessert, and frequency of consumption. This data will be used for market analysis and customer behavior studies.